avances quintos

+++ src/controllers/cart.controllers.js
@@ -1,38 +1,84 @@
 const catchError = require('../utils/catchError');
 const Cart = require('../models/Cart');
+const Product = require('../models/Product');
+const Category = require('../models/Category');
 
 const getAll = catchError(async(req, res) => {
-    const results = await Cart.findAll();
+    const userId =req.user.id;
+    const results = await Cart.findAll({
+        where: {userId},
+        include: [
+            {
+                model: Product,
+                attributes:{exclude: ['createdAt', 'updatedAt']},
+                include:[
+                    {
+                        model: Category,
+                        attributes: ["name"]
+                    }
+                ]
+            }
+        ]
+    });
     return res.json(results);
 });
 
 const create = catchError(async(req, res) => {
-    const result = await Cart.create(req.body);
+    /*
+    {
+    "quantity": 3,//!cliente
+    "productId":3,//!cliente
+    "userId": 3//?backend
+    } 
+    */
+   const userId = req.user.id;
+   const {quantity, productId} = req.body;
+
+
+    const result = await Cart.create({userId, quantity, productId});
     return res.status(201).json(result);
 });
 
 const getOne = catchError(async(req, res) => {
     const { id } = req.params;
-    const result = await Cart.findByPk(id);
+    const userId = req.user.id;
+    const result = await Cart.findByPk(id, {
+        where: {userId},
+        include: [
+            {
+                model: Product,
+                attributes:{exclude: ['createdAt', 'updatedAt']},
+                include:[
+                    {
+                        model: Category,
+                        attributes: ["name"]
+                    }
+                ]
+            }
+        ]
+    });
     if(!result) return res.sendStatus(404);
     return res.json(result);
 });
 
 const remove = catchError(async(req, res) => {
     const { id } = req.params;
-    const result = await Cart.destroy({ where: {id} });
+    const userId = req.user.id;
+    const result = await Cart.destroy({ where: {id, userId} });
     if(!result) return res.sendStatus(404);
     return res.sendStatus(204);
 });
 
 const update = catchError(async(req, res) => {
     const { id } = req.params;
+    const userId = req.user.id;
 
-    
+    delete res.body.userId;
+    delete  res.body.productId;
 
     const result = await Cart.update(
         req.body,
-        { where: {id}, returning: true }
+        { where: {id, userId}, returning: true }
     );
     if(result[0] === 0) return res.sendStatus(404);
     return res.json(result[1][0]);
